Accept fractional Actual Rows in EXPLAIN JSON
Amp-Thread-ID: https://ampcode.com/threads/T-019feecc-d1b8-751e-bf47-61ef3019d512

+++ logs/querysample/normalize.go
@@ -23,7 +23,7 @@ type planNodeGroupingSet struct {
 
 type planNode struct {
 	ActualLoops                 *int64                   `json:"Actual Loops,omitempty"`
-	ActualRows                  *int64                   `json:"Actual Rows,omitempty"`
+	ActualRows                  *float64                 `json:"Actual Rows,omitempty"` // PostgreSQL 18+ reports fractional rows
 	ActualStartupTime           *float64                 `json:"Actual Startup Time,omitempty"`
 	ActualTotalTime             *float64                 `json:"Actual Total Time,omitempty"`
 	Alias                       *string                  `json:"Alias,omitempty"` // does not need normalize (table alias as specified in the query)
